+++ app/src/application/mod.rs
@@ -1,4 +1,5 @@
 pub mod client;
 pub mod server;
+pub mod standalone;
 pub mod types;
 pub mod utils;
