+++ app/src/hyprland/types/socket_type.rs
@@ -1,0 +1,4 @@
+pub enum HyprSocketType {
+    Event,
+    Command,
+}
