+++ app/src/hyprland/utils.rs
@@ -1,7 +1,6 @@
 use super::types::{HyprEventList, HyprSocketType};
-use crate::{ipc::*, HyprvisorResult};
+use crate::{global::BUFFER_SIZE, ipc::*, HyprvisorResult};
 
-use core::panic;
 use std::env;
 use tokio::{io::AsyncReadExt, net::UnixStream};
 
@@ -9,10 +8,7 @@ pub const HYPRLAND_SOCKET_CONNECT_ATTEMPT: u8 = 3;
 pub const HYPRLAND_SOCKET_CONNECT_DELAY: u64 = 100;
 
 pub(super) fn hyprland_socket(socket_type: &HyprSocketType) -> String {
-    let instance_signature = match env::var("HYPRLAND_INSTANCE_SIGNATURE") {
-        Ok(var) => var,
-        Err(_) => panic!("Is hyprland running?"),
-    };
+    let instance_signature = env::var("HYPRLAND_INSTANCE_SIGNATURE").expect("Is Hyprland running?");
 
     let socket_name = match socket_type {
         HyprSocketType::Command => ".socket.sock",
@@ -24,8 +20,10 @@ pub(super) fn hyprland_socket(socket_type: &HyprSocketType) -> String {
         .unwrap_or_else(|_| format!("/tmp/hypr/{instance_signature}/{socket_name}"))
 }
 
-pub(super) async fn send_hyprland_command(cmd: &str) -> HyprvisorResult<Vec<u8>> {
-    let mut buffer = vec![0; 4096];
+pub(super) async fn send_hyprland_command(command: &str) -> HyprvisorResult<Vec<u8>> {
+    log::debug!("send_hyprland_command: {}", command);
+
+    let mut buffer = vec![0; *BUFFER_SIZE];
 
     connect_to_socket(
         &hyprland_socket(&HyprSocketType::Command),
@@ -33,7 +31,7 @@ pub(super) async fn send_hyprland_command(cmd: &str) -> HyprvisorResult<Vec<u8>>
         HYPRLAND_SOCKET_CONNECT_DELAY,
     )
     .await?
-    .try_send_and_receive_bytes(cmd.as_bytes(), &mut buffer, 10)
+    .try_send_and_receive_bytes(command.as_bytes(), &mut buffer, 10)
     .await
     .map(|recv_len| buffer[..recv_len].to_vec())
 }
@@ -42,10 +40,11 @@ pub(super) async fn fetch_hyprland_event(
     stream: &mut UnixStream,
     buffer: &mut [u8],
 ) -> HyprEventList {
-    match stream.read(buffer).await {
-        Ok(bytes) if bytes > 0 => buffer[..bytes].into(),
-        Ok(_) | Err(_) => {
-            panic!("Connection closed from Hyprland event socket");
-        }
-    }
+    log::debug!("fetch_hyprland_event");
+
+    stream
+        .read(buffer)
+        .await
+        .map(|len| buffer[..len].into())
+        .expect("Connection closed from Hyprland event socket.")
 }
